+++ code/packages/ruby/conduit/lib/coding_adventures_conduit.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 require_relative "coding_adventures/conduit/version"
+require_relative "coding_adventures/conduit/halt_error"
+require_relative "coding_adventures/conduit/handler_context"
 require_relative "coding_adventures/conduit/request"
 require_relative "coding_adventures/conduit/route"
 require_relative "coding_adventures/conduit/router"